// src/lib/constants/api.ts
export enum ApiEndpoints {
  TOOLS = 'tools',
  IMPORT_URL = 'import/url',
  IMPORT_SCHEMA = 'import/content',

  CONFIGURATIONS = 'configurations',
  BY_SCHEMAS = 'by-schemas',
  BY_TYPE = 'by-type',
  PROJECTS = 'projects',
  ACTIONS = 'actions',

  WORKFLOWS = 'workflows',

  WORKFLOW = 'workflow',
  WORKFLOW_CHANGE = 'workflow-change',
  WORKFLOW_CHANGE_PROCESS_INSTANCE = 'workflow-change/process-instance',
  WORKFLOW_CHANGE_DEFINITION_ID = 'workflow-change/definition-id',

  EXECUTE = 'execute',
  DEPLOY = 'deploy',

  DEPLOYMENTS = 'deployments',
  COMPOSE = 'compose',
}